更新提交脚本支持PG 11

+++ site.docbuilder/home/pgdoc/gitrep/builddoc.sh
@@ -1,11 +1,29 @@
 #!/bin/bash
 
+if [ ! $# -eq 1 ]
+then
+ echo "Usage: $0 pgversion"
+ exit -1
+fi
+
+PGVERSION=$1
+
+if [ ! -d ../postgresql-${PGVERSION} ]
+then
+ echo "could not find ../../postgresql-${PGVERSION}"
+ exit -1
+fi
+
+rm -f ./pgdoc-cn/build
+ln -sf ../../postgresql-${PGVERSION} ./pgdoc-cn/build
+cp -rf ./pgdoc-cn/postgresql/doc/src/sgml/* ./pgdoc-cn/build/doc/src/sgml/
+
 rm -rf html_out
 rm -rf pgdoc-cn/tools/html_out
 
 cd pgdoc-cn
 
-perl tools/encoding_convert.pl
+# perl tools/encoding_convert.pl
 cd build/doc
 echo sed -e 's/\$\{updatedate\}/'`date +%Y-%m-%d`'/' -i src/sgml/pgdoccn-notes.sgml|sh
 gmake html
